@@ -1,17 +1,18 @@
-import { Injectable } from '@angular/core';
-import { CanActivate, Router } from '@angular/router';
-import { AuthService } from '../services/auth.service';
+import { Injectable } from "@angular/core";
+import { CanActivate, Router } from "@angular/router";
+import { AuthService } from "../services/auth.service";
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: "root"
 })
 export class AuthGuardService implements CanActivate {
-
-  constructor(private authService: AuthService, private router: Router) { }
+  constructor(private authService: AuthService, private router: Router) {}
 
   canActivate() {
-    if (this.authService.isLoggedIn) return true;
-    this.router.navigate(['/login']);
+    if (this.authService.currentUser) {
+      return true;
+    }
+    this.router.navigate(["/auth/login"]);
     return false;
   }
 }
